refactor(portal): relocate v3-needed code out of the v2 tree (v2 removal, step 1)
Pure relocations so the v2 portal tree can later be deleted without
touching live code paths:

- Move apps.gql/apps.generated (InitialApp) from the v2 Apps page into
  the PortalV3 Team page tree; repoint both consumers and both test
  mocks. Codegen verified idempotent at the new path.
- Repoint api/mcp's four AppStore imports to the PortalV3 twins
  (diff-verified: form-schema/types/dataTransforms identical, utils and
  AppStoreFormTypes strict supersets).
- Move the public kiosk wrapper scenes/Portal/Kiosk -> scenes/PortalV3/Kiosk
  unchanged; its relative ActiveKiosk import now resolves to the v3 copy
  (byte-identical except useLegacyKioskRequest, where the v3 copy keeps
  the [appId, requestVersion] memo deps that #2059 dropped from v2 during
  a lib/rp extraction).
