+++ web/tests/unit/portal-v3-team-overview.test.tsx
@@ -23,7 +23,7 @@ jest.mock("@/api/helpers/graphql", () => ({
 
 const InitialApp = jest.fn();
 jest.mock(
-  "@/scenes/Portal/Teams/TeamId/Apps/page/graphql/server/apps.generated",
+  "@/scenes/PortalV3/Teams/TeamId/Team/page/graphql/server/apps.generated",
   () => ({
     getSdk: () => ({ InitialApp }),
   }),
